Corrected AUTHOR keyword description in map FITS file.
The AUTHOR keyword is not meant to describe who created the FITS file
in which it resides.  Rather it is intended to be to identify who
compiled the original information and data, such as from a
publication, found in the FITS file.  The CREATOR keyword, such as the
one that was introduced in an earlier change, should instead be used
to described who created the FITS file.

See the FITS standard dictionary for additional details on the AUTHOR
keyword:
    https://heasarc.gsfc.nasa.gov/docs/fcg/standard_dict.html