Implemented improved "is zero" check.  (#27)
* Added a MaRC::almost_zero() function.

The newly introduced MaRC::almost_zero() function should be used to
check if a floating point number can be considered zero within a
certain degree of precision.  So far only MaRC::MosaicImage will
leverage this function.

* Document bug in floating point Matrix/Vector equality.

The MaRC Matrix and Vector equality operator==() as implemented is
only reliable for integer element types.  It is not as accurate for
floating point element types since the elements are compared for exact
equality.  Document this bug for now.  It doesn't impact MaRC yet
since MaRC only uses the Matrix and Vector equality operators in the
corresponding unit tests, and only for integers in those cases.