WIP implementation of source image data auto-scaling.
The goal of this work-in-progress implementation is to automatically
determine the best source image data scale and offset parameters to
retain as many significant digits as possible when storing that data
in integer typed maps.  It is is currently intended for use with MaRC
virtual images, such as cosine of emission angle images
(MaRC::MuImage), latitude images (MaRC::LatitudeImage), etc.

Automatically determining the best scale and offset allow removal of
the hard-coded scale and offset values in MuImage, Mu0Image and
CosPhaseImage implementations.