Reduce number of MapFactory<> parameters.
In preparation of supporting auto-scaling of data, reduce the number
of parameters passed to the MapFactory<>::make_map(),
MapFactory<>::plot_map() and MapFactory<>::plot() methods
significantly.  They removed parameters will be instead be moved to a
structure where they can be passed in efficiently as a single
parameter.

This is still a work in progress.