PhotoImageFactory: Fixed flat-field correction.
An incorrect 'fitsfile' pointer was passed to the CFITSIO function
that read the flat field image FITS file, resulting in corruption.
Use the correct pointer.  The flat fielding code was also moved to a
separate function to prevent this sort of thing from happening, as
well as to improve the overall quality of the code.