Latitude/longitude visibility improvements (#38)
* Determine visibility by verifying the emission and incidence (only
checked if configured to do so) angles are positive.  Greatly
simplifies this method, correctly takes into account necessary
viewing geometry variables, and removes requirement that the body is
modeled as an oblate spheroid.

Somehow this change was lost back in 2005, soon after the v0.9.7
release.

* PhotoImage: Return boolean result directly rather than by 'if(true) return true'

* Latitude/longitide visibility improvements.

Improved latitude/longitude visibility checks in the MuImage, Mu0Image
and CosPhaseImage classes by leveraging emission and incidence angles
where appropriate.  The new implementations are much simpler, take
into account observer range, and no longer require that the observed
body is an oblate spheroid.

* PhotoImage: Fix off-by-one-half error in pixel coordinates

Do not round calculated floating point image line and sample.  They
are in 'pixel coordinates'.  In 'pixel coordinates', 0..1 is inside
pixel 0, 1..2 is inside pixel 1, etc.

* Make 360 degree longitude range clamping more robust.

Implement 360 longitude range clamping in terms of std::fmod() instead
of simple +/- 360 conditionals.  This handles cases when longitudes
are outside of the +/-360 range.

* Added OblateSpheroid test.

* OblateSpheroid: Simplify centric radius calculation.

* Improved Doxygen documentation.

* Added MaRC::Geometry unit test.

The new Geometry_Test unit test verifies that all functions in the
MaRC::Geometry namespace work as expected.  This is particularly
important since coordinate system rotations are a critical part of
MaRC.

* Do not include <iostream> unnecessarily.

* Geometry_Test: Loosen 'almost equal' criterion.

Increase the ULPs from 2 to 4 to loosen the "almost equal" criterion
for determining if two floating numbers can be considered equal.  2
ULPs was too strict for a 32 bit platform with apparently poor
trigonometric function implementations.

* PhotoImage: Prefer hypot() over sqrt() for pixel range calc.

* OblateSpheroid_Test: Added simple latitude test.