PhotoImage: Fixed sky removal based averaging weight.
The averaging weight based on the distance of the sky to the left of a
point on the body in the image was incorrect.  A scan on the first
line of image was always performed instead of the line corresponding
to the latitude and longitude of the data being read.  This issue only
occurred when both sky removal and weighted averaging of mosaiced
images were enabled.  Note that the correct data was still read. Only
the data's averaging weight was potentially off.

Generation of the sky removal mask was also slightly optimized.