Simplified MU, MU0 and PHASE images.
Dropped the visibility checks for the MuImage, Mu0Image and
CosPhaseImage virtual images.  They were redundant.  There was also no
need to check if the data values were positive (i.e. angle in the
interval [-90, 90]) since negative values (i.e. [-1, 0)) are perfectly
valid as well.  Doing so improved speed by up to 100%.

The existence of CosPhaseImage data is also no longer tied to a
positive cosine of the emission of the angle (mu), meaning cosine of
phase angle data will be plotted in areas of the map where mu < 0.