VirtualImage: Correct automatic offset calculation.
The offset calculated for LONGITUDE map planes was always zero, which
was incorrect. MaRC uses a longitude data range of [0, 360], not [-180,
180].  The scale for LONGITUDE maps was calculated correctly but the
offset was not. Assuming a map with 16 bit signed integer data was
being plotted, the maximum value is 32767. The LONGITUDE map data was
be scaled by 100 with an offset of 0, making the maximum plotted data
value 36000, which is larger than the largest 16 bit signed integer
32767. The result was longitude values greater than 327 degrees ended
up being dropped from the map. The offset value should have been
-18000 to shift the map value into the chosen integer range of values,
i.e. [0, 360] scaled by 100 and offset by -18000 results in map data
in the range [-18000, 18000] which is well within the 16 bit signed
integer data range [-32768, 32767]. In this case, the corresponding
FITS BSCALE and BZERO values needed to convert from the map value to
the physical value could be 0.01 and 180, respectively, which is now
what is calculated.

This problem only occured when creating a LONGITUDE integer typed
map plane containing scaled longitude values larger than a certain
value (e.g. larger than 32767 in 16 bit signed integer maps). Floating
point maps didn't exhibit this problem.

A new Scale_Offset_Test unit test has been added to confirm the
correct behavior.

(Fixes #42)