LongitudeImage: Do not force into 360 degree range.
The LongitudeImage virtual image no longer attempts to force the given
longitude into the 360 degree range.  It simply converts the given
longitude in radians to degrees.  It is up to the caller to decide
what the longitude range should be.

No change in behavior will be seen for those that use the MaRC
supplied maps since they always supplied longitudes in the [0, 2pi]
radians, i.e. [0, 360] degrees, range.