Address LongitudeImage test failure.
The LongitudeImage_Test unit test failed on a 32 bit virtual machine
due to lack of precision of two values that were basically zero for
the purposes of this test, i.e. 3.747e-15 and the machine epsilon() of
2.2204e-16, didn't compare as equal.  In this case check if the values
are both almost zero instead since the "almost equal" implementation
isn't suitable for comparison of values that are nearly zero.