Make map projections non-template classes.
Map projections were implemented as class templates to support
different map data types. However, the bulk of the map operations and
calculations are type independent. Rather than implement the entire
projection as a class template, rely on member function templates
to instead isolate type-specific operations.  This greatly reduces the
static footprint of the marc binary.  For example, including both the
MaRC library and program, the static footprint was reduced by 28% for
a 64-bit Linux build with g++ 7.2.

Furthermore, the MaRC API ends up being a bit simpler.  Code
duplication in the the MaRC input file parser, for example, was
reduced.

Closes #34.