Incremental merge of multiple changes (#64)
* Clean up and refactor root finding code.

* Orthographic: Coding style clean-up.

* Implement minimal root finding test.

* Only enable arch tuning flag detection on request.

* ViewingGeometry: Fixed uninitialized class member.

Add missing initialization of 'use_terminator_' member in base member
initializer list.  Corrects a problem where the terminator could be
used unexpectedly when determining visibility of a point on the body.

This is regression was introduced when the "flags" bit mask was
removed a while back.

* Address LongitudeImage test failure.

The LongitudeImage_Test unit test failed on a 32 bit virtual machine
due to lack of precision of two values that were basically zero for
the purposes of this test, i.e. 3.747e-15 and the machine epsilon() of
2.2204e-16, didn't compare as equal.  In this case check if the values
are both almost zero instead since the "almost equal" implementation
isn't suitable for comparison of values that are nearly zero.

* Correct thrown exceptions.

Made exceptions better match the conditions under which they are
thrown.  For example, throw std::invalid_argument instead of
std::range_error where appropriate.

* Mercator: Begin fixing incorrect 90 deg max lat.

* Handle potential overflow in scale/offset calc.

* Mercator: Minor consistency update.

* PolarStereographic: Minor code optimization.

* Make map projections non-template classes.

Map projections were implemented as class templates to support
different map data types. However, the bulk of the map operations and
calculations are type independent. Rather than implement the entire
projection as a class template, rely on member function templates
to instead isolate type-specific operations.  This greatly reduces the
static footprint of the marc binary.  For example, including both the
MaRC library and program, the static footprint was reduced by 28% for
a 64-bit Linux build with g++ 7.2.

Furthermore, the MaRC API ends up being a bit simpler.  Code
duplication in the the MaRC input file parser, for example, was
reduced.

Closes #34 

* Update copyright years for 2018 modified files.

* Removed unnecessary NDEBUG autoheader template.

The AX_CHECK_ENABLE_DEBUG macro already provides the necessary
autoheader template for the NDEBUG preprocessor symbol.  No need to
explicitly provide one in configure.ac.

* Mathematics: Added note about undesirable config.h include.

* Removed unnecessary MaRC/config.h include.

* Work around buggy std::bind().

Some older implementations of std::bind() (e.g. in libc++) were unable
to handle binding a const member function, such as the underlying map
equation functions in the Mercator and Polar Stereographic projection
implementations.  Work around that problem by moving the functions
being bound to the file scope.

* configure: Bump MaRC version to 0.9.10-beta.