Correct problems in conformal maps (#71)
* Added Mercator and Polar Stereographic test input files.

* Fixed seg fault in Polar Stereographic projection.

* Added Polar Stereographic projection test.

* OblateSpheroid: Removed flattening parameter.

The OblateSpheroid class now requires both the equatorial and polar
radii be supplied.  The flattening was not needed after
initialization.

* Mercator: Corrected buffer overflow.

The Mercator grid creation routine computed grid image array offsets
for longitude lines incorrectly, i.e. "[k * lines + i]" instead of the
correct "[k * samples + i]", where and "k" and "i" are the line and
sample in the grid image being written to.

* tests: Increase grid interval in test maps.

* tests: Clean up generated maps.

* MosaicImage: Correct spelling error in comment.

* root_find: Pass bracket values instead of guess.

Rather than passing in an initial guess to MaRC::root_find(), pass a
pair of bracket values between a search for the root will be
performed.  Allows for a more reliable search.  Fixes #67 

* Corrected Polar Stereographic equation regression.

* Take into account which pole is at the center when calculating the
  Polar Stereographic scale distortion, as well as when creating the
  grid.

* Improve MaRC::root_find() documentation.

* Added basic Mercator projection test.

* MaRC_input: Corrected conformal map documentation.

* Removed incorrect references to "Transverse Mercator".  This
  implementation is a regular Mercator projection, not transverse.

* Removed misguided attempts at allowing the user to set the maximum
  latitude.  To be revisited.

* NEWS: Updated notes about conformal maps.