configure.ac: Remove "foreign" Automake option.
Remove the "foreign" Automake option so that the generic INSTALL file
is included in the generated source distribution.  It was originally
included to prevent Automake from copying the GPLv3 COPYING file to
the top-lvel source directory when adding missing files, but that
really isn't necessary since Automake won't overwrite an existing
COPYING file unless the "--force" command line option is used.