Fixed mosaic averaging regression.
A regression was introduced when C++11 rvalue references were
introduced to MaRC that caused weighted and unweighted averaging to be
disabled if either was selected.  MaRC automatically disables
averaging if only one image (photo) exists in the map plane, but in
this case the number of images was based on the size of a moved (via
std::move()) list.  That resulted in the number of images in the
mosaic incorrectly assumed to be zero.  Fixes #80.