Further expanded code coverage. (#81)
* Fixed mosaic averaging regression.

A regression was introduced when C++11 rvalue references were
introduced to MaRC that caused weighted and unweighted averaging to be
disabled if either was selected.  MaRC automatically disables
averaging if only one image (photo) exists in the map plane, but in
this case the number of images was based on the size of a moved (via
std::move()) list.  That resulted in the number of images in the
mosaic incorrectly assumed to be zero.  Fixes #80.

* tests: Expand MaRC feature use in test input file.

Code coverage results revealed large portions of MaRC that were not
exercised by its unit tests.  Additional test maps were added to the
MaRC test input file `test_map.marc' to force most of the uncovered
code to run during test runs.

* parse.yy: Set range before arcseconds per pixel.

The MaRC library currently requires the range to be set prior to the
number of arcseconds per pixel.  Adjust the code in the parser grammar
accordingly.

* PhotoImage: Fix incorrect assertions.

Two assert() calls were being triggered due to an correct check for a
"begin" iterator being less than the result of std::find().  It is
possible for std::find() to validly return the begin iterator.  Allow
the begin iterator to equal the std::find() result as well.