Notify progress observers of map completion.
Some map projections, such the Orthographic projection, will only plot
a portion of the map, in which case the underlying plot count in the
progress notifier will never reach the map size, meaning subscribed
observers may not know that mapping was completed.  Explicitly notify
all observers that mapping is done by passing a plot count equal to
the map_size.