src: Document bug in parser error reporting.
The reported line number is off when the parser explicitly calls
yyerror().  For example, if a negative KM_PER_PIXEL value is set in an
input file the reported line is actually the first line that isn't
solely whitespace after the line containing the invalid KM_PER_PIXEL
value.  Line numbers in syntax errors, on the other hand, are correct.