lib: Corrected auto data scaling comments.
Clarified and corrected comments regarding calculation of scale and
offset values for MaRC virtual images.  The comments incorrectly
implied that no loss of significant digits would occur when assigning
the scaled and offset floating point values to an integer.  Some loss
of significant digits will always occur.  The goal is to retain as
many significant digits as possible without a complete loss of them.