lib: Fix problem single longitude being mapped.
Corrected a problem where MaRC's Simple Cylindrical projection could
not handle the case where lower and upper longitudes in the map were
the same, signifying a 360 degree range. For example, passing 0 as the
lo_lon and hi_lon longitude values in the MaRC::SimpleCylindrical
constructor results in a map with a single longitude 0 being mapped
across all map pixels.  Fixes #95.