Don't bother storing std::string in optional<>.
It's redundant to store a std::string in a MaRC::optional<> since we
can just rely on the std::string::empty() method to determine if a
string has a value in this particular context.  Once MaRC switches to
C++17, we may want to revisit this decision since MaRC's simple C++14
based implementation::optional<> is not as a good as C++17's
std::optional<> in terms of construction and destruction performance.