MapFactory_t: Corrected blank value range check.
The check that attempted to verify that the user supplied map blank
value fit within the map data type range always succeeded since the
user supplied value was explicitly cast to the map data type before
the check.  Cast to the map data type after the check to allow the
test to fail when it should.