tests: Orthographic projection test now passes.
The orthographic projection test was failing due to use of a map data
type that could not provide the precision necessary to support the
floating point equality strictness (small ulp value) in the test.
Force the use of either a floating point map data type, or an integer
type that is at least 64 bits wide.