lib: Drop zero check in MaRC::quadratic_roots().
Drop the newly added "b == 0" check in MaRC::quadratic_roots() in
favor of a branchless implementation that relies on a newly added
MaRC::sgn() function.  The new MaRC::sgn() function returns the sign
of a real number.  It differs from the signum() function in that it
interprets the value zero as having a positive sign, i.e. 1, instead
of 0.  This behavior is required for the numerically stable form of
the quadratic formula implemented in MaRC::quadratic_roots().