src: Move input file number conversion to parser.
Handle MaRC input file related string to number conversion in the
parser to make it possible to convert 64 bit integers in string form
directly to the underlying integer type instead of using variable of
type double as an intermediary.  Addresses an issue where the smallest
and largest 64 bit integers could be not be passed from the scanner to
the parser since the largest integer that can be stored in a 64 bit
IEEE 754 floating point value (type double) without loss of precision
is +/- 2^53.