lib: Add MaRC::data() function.
As MaRC moves to supporting both data arrays and containers with
array-like interfaces, the need to obtain the data contained in them
in a simple manner occurred.  Support for the MaRC::data() function
was added.  It implements the C++17 std::data() function semantics,
and provides a uniform way to get a pointer to the underlying block of
memory in the array or container.