FITS_image: Make DATAMIN/MAX write type neutral.
Make the FITS_image datamin() and datamax() members to be member
templates instead of just assuming that the type 'double' is
sufficient in all cases.  This allows the DATAMIN and DATAMAX values
to match the map data type, which makes sure these values actually
match the minimum and maximum values in the map.