lib: Factor our user-specified extrema.
Make user-specified extrema a separate parameter to
MapFactory<>::make_map() so that a single plot_info<> instance may be
used between multiple calls.  This is useful for multi-plane maps
where the map dimensions do not change, as well as for tracking the
physical data minimum and maximum between all map planes.