Use a single plot_info<> instance per map plane.
There was no need to create a new set of plot_info<> and
Progress::Console instances for each map plane.  A single instance of
each is sufficient.  This allows physical data extrema for a given
multi-plane FITS image (MaRC generated map) to be automatically
tracked, and written to the map FITS file in the DATAMIN and DATAMAX
FITS cards.