src: Do not set map DATAMIN/MAX from source img.
Do not set the map DATAMIN and DATAMAX values through the source image
factories.  Instead, rearrange the code so that those values, if
available, end up being used when generating the map.  This prevents
user-specified min/max values from being overridden, as well as
potential mismatches with the physical data min/max for data that was
actually mapped.