extrema: Store minimum and maximum as optional<>.
Store MaRC::extrema<> minimum and maximum values as instances
MaRC::optional<> to make it easier to determine if a either was set.
This is primarily useful for the case where the user sets one of
DATAMIN or DATAMAX in their MaRC input file.

This change also addresses a regression where it was no longer
possible to only set either DATAMIN or DATAMAX, but not both.  The
resulting MaRC::extrema<> instance would not contain valid set of extrema.