Automatically populate map parameters (#113)
* lib: Remove unit() accessors.

Move all physical data unit handling to the MaRC program, where it is
actually used.

* src: Populate parameters through image factories.

The SourceImageFactory objects that generate the data being mapped
may contain information that may be used to automatically populate the
map parameters.  Implement a new populate_parameters() in each
concrete SourceImageFactory sub-class that does so.

* src: Attempt to populate map parameters.

Attempt to automatically populate map parameters from the
SourceImageFactory objects used to create the data to be plotted.

* src: Prevent automatic decrease in data type size.

When automatically populate map parameters, prevent a decrease in the
map data type size when iterating over the map plane image factories.

* src: Allow user to override map parameters.

Avoid overriding map parameters explicitly set by the user.

* MapParameters: Remove unused NaN code.

* src: Fix Doxygen warnings.

* SimpleCylindrical: Minor optimization.

Only calculate the longitude-to-sample ratio once rather than each
time a longitude is calculated.

* src: Spelling correction.

* Mathematics: Correct spelling.

* Create convenience library for testing purposes.

* Add initial MapParameters unit test.

* src: Prefer 'override' over 'virtual'.

* Remove remaining uses of std::*stream.

* MapFactory_t: Prefer a lambda over std::bind().

* lib: Pass the plot() functor by reference.

Improve performance by passing the MaRC::MapFactory::plot_type functor
by reference instead of by value since the underlying std::function<>
class template has several fields.

* Clarify MaRC::almost_zero() second argument names.

The second argument to MaRC::almost_zero() is the number of floating
point epsilons, not units in the last place (ulps).

* lib: Track min and max values of mapped data.

Track minimum and maximum values of physical data that was actually
mapped so that they are readily available to the caller of
MaRC::MapFactory::make_map().

* src: Automatically write DATA{MIN,MAX} values.

Write the minimum and maximum values of physical data that was
actually mapped as the values of the DATAMIN and DATAMAX FITS keys,
respectively, in the FITS header of the MaRC generated map file.

The values will not exceed user-specified minimum and maximum physical
data values.

* FITS_image: Write file creation date early on.

* FITS_image: Make DATAMIN/MAX write type neutral.

Make the FITS_image datamin() and datamax() members to be member
templates instead of just assuming that the type 'double' is
sufficient in all cases.  This allows the DATAMIN and DATAMAX values
to match the map data type, which makes sure these values actually
match the minimum and maximum values in the map.

* plot_info: Make minimum/maximum type match map.

Change the MaRC::plot_info class into a class template so that it
becomes possible to make the minimum and maximum field types match the
map data type.  This provides allows the minimum and maximum values
to always match the corresponding values in the map.

* extrema: Refactor extrema handling to new class.

* Use a single plot_info<> instance per map plane.

There was no need to create a new set of plot_info<> and
Progress::Console instances for each map plane.  A single instance of
each is sufficient.  This allows physical data extrema for a given
multi-plane FITS image (MaRC generated map) to be automatically
tracked, and written to the map FITS file in the DATAMIN and DATAMAX
FITS cards.

* Display mismatched sizes on map write error.

* SourceImageFactory: Use new MaRC::extrema<> class.

Leverage the new MaRC::extrema<> class template to make minimum and
maximum physical data value handling more robust.

* Begin switching to underscore naming convention.

MaRC is transitioning from PascalCase to underscore based name
delimiting, such as what is used in the standard C++ library.  Start
with MapParameters -> map_parameters since it is being actively
developed.

* Fix some 'clang-tidy' warnings.