lib: Make data weight double instead of size_t.
Make the physical data weight type used for weighted averaging when
creating a mosaic a type double instead of std::size_t.  Don't assume
std::size_t is suitable as the data weight type for all images that
may be composited.

Remove comments about data weights being related to the edge of an
image.  There is no need to assume or expect that all images will base
data weights on such a criterion.